Drafting LineTAP registration.
This is based on what I've written for EPN-TAP.

I've not actually filled the example record, as I intend to take that
from live TOSS once I've make the prototype.
Drafting LineTAP registration.

This is based on what I've written for EPN-TAP.

I've not actually filled the example record, as I intend to take that
from live TOSS once I've make the prototype.