Adding a registry record for linetap
This is now necessary because that's where our "utype" lives in.

While doing this, gently streamlined the metadata a bit: dumped
the "protocol" from title (one could argue it should be there, but
RegTAP has survived without; anyway, if it's in, it needs to go in a
bit differently).

I'm also adding myself as an author.