Adding element as a linetap column.
This is because InChI tooling is quite a bit weaker than thought (e.g.,
there is no pure-Java implementation of it).  Hence, we should only
require the use on InChIs when we actually deal with molecules (where
there's little in terms of sane alternatives).

Also, I'm un-requiring inchi and inchikey except for molecules, and I'm
requiring element for atoms.  Reflecting that in the columns table would
admittedly be nice.